Enhance CUDA API usage and memory pooling
Refactors core components in the Pesti runner to utilize direct CUDA
driver APIs rather than high-level wrappers. This includes migrating
async transfers and updating benchmark examples for improved precision
and lower-level control.

Additionally, memory pool support is expanded:
- Increased size classes up to 16MB to handle larger tensors
  efficiently.
- Improved allocation logic allows fallbacks to direct device allocation
  when pooled sizes do not match required memory.

+++ pesti-conformance/src/lib.rs
@@ -244,7 +244,7 @@ fn run_pesti_inference(model_path: &Path) -> Result<String> {
 
     // Initialize token embeddings from loaded weights
     let embed_dim = model.embed_dim;
-    let _vocab_size = model.vocab_size as usize;
+    let _vocab_size = model.vocab_size;
 
     // Create a simple input: single token [0] for conformance test
     let input_tokens: Vec<i32> = vec![0i32];
